+++ openedx_events/event_bus/tests/test_loader.py
@@ -9,7 +9,7 @@ from unittest import TestCase
 from django.test import override_settings
 
 from openedx_events.data import EventsMetadata
-from openedx_events.event_bus import _try_load, get_producer
+from openedx_events.event_bus import _try_load, get_producer, make_single_consumer
 from openedx_events.learning.signals import SESSION_LOGIN_COMPLETED
 
 
@@ -30,7 +30,7 @@ class TestLoader(TestCase):
     def test_unconfigured(self):
         with assert_warnings(["Event Bus setting DOES_NOT_EXIST is missing; component will be inactive"]):
             loaded = _try_load(
-                setting_name="DOES_NOT_EXIST",
+                setting_name="DOES_NOT_EXIST", args=(1), kwargs={'2': 3},
                 expected_class=dict, default={'def': 'ault'},
             )
         assert loaded == {'def': 'ault'}
@@ -39,10 +39,10 @@ class TestLoader(TestCase):
     def test_success(self):
         with assert_warnings([]):
             loaded = _try_load(
-                setting_name="EB_LOAD_PATH",
+                setting_name="EB_LOAD_PATH", args=(), kwargs={'2': 3},
                 expected_class=dict, default={'def': 'ault'},
             )
-        assert loaded == {}
+        assert loaded == {'2': 3}
 
     @override_settings(EB_LOAD_PATH='builtins.list')
     def test_wrong_type(self):
@@ -51,7 +51,7 @@ class TestLoader(TestCase):
                 "component will be inactive"
         ]):
             loaded = _try_load(
-                setting_name="EB_LOAD_PATH",
+                setting_name="EB_LOAD_PATH", args=([1, 2, 3],), kwargs={},
                 expected_class=dict, default={'def': 'ault'},
             )
         assert loaded == {'def': 'ault'}
@@ -64,7 +64,7 @@ class TestLoader(TestCase):
                 "component will be inactive"
         ]):
             loaded = _try_load(
-                setting_name="EB_LOAD_PATH",
+                setting_name="EB_LOAD_PATH", args=(1), kwargs={'2': 3},
                 expected_class=dict, default={'def': 'ault'},
             )
         assert loaded == {'def': 'ault'}
@@ -77,20 +77,20 @@ class TestLoader(TestCase):
                 "component will be inactive"
         ]):
             loaded = _try_load(
-                setting_name="EB_LOAD_PATH",
+                setting_name="EB_LOAD_PATH", args=(1), kwargs={'2': 3},
                 expected_class=dict, default={'def': 'ault'},
             )
         assert loaded == {'def': 'ault'}
 
-    @override_settings(EB_LOAD_PATH='builtins.len')
+    @override_settings(EB_LOAD_PATH='builtins.dict')
     def test_bad_args_for_callable(self):
         with assert_warnings([
                 "Failed to load <class 'dict'> from setting EB_LOAD_PATH: "
-                "TypeError('len() takes exactly one argument (0 given)'); "
+                "TypeError('type object argument after * must be an iterable, not int'); "
                 "component will be inactive"
         ]):
             loaded = _try_load(
-                setting_name="EB_LOAD_PATH",
+                setting_name="EB_LOAD_PATH", args=(1), kwargs={'2': 3},
                 expected_class=dict, default={'def': 'ault'},
             )
         assert loaded == {'def': 'ault'}
@@ -112,3 +112,17 @@ class TestProducer(TestCase):
                 event_key_field='user.id', event_data={},
                 event_metadata=EventsMetadata(event_type='eh')
             ) is None
+
+
+class TestConsumer(TestCase):
+
+    @override_settings(EVENT_BUS_CONSUMER=None)
+    def test_default_does_nothing(self):
+        """
+        Test that the default is of the right class but does nothing.
+        """
+        consumer = make_single_consumer(topic="test", group_id="test", signal=SESSION_LOGIN_COMPLETED)
+
+        with assert_warnings([]):
+            # Nothing thrown, no warnings.
+            assert consumer.consume_indefinitely() is None